+++ CHANGELOG.md
@@ -172,6 +172,27 @@
   `_`.
   ([Jiangda Wang](https://github.com/frank-iii))
 
+- The language server can now suggest a code action to import modules
+  for existing code which references unimported modules:
+
+  ```gleam
+  pub fn main() {
+    io.println("Hello, world!")
+  }
+  ```
+
+  Becomes:
+
+  ```gleam
+  import gleam/io
+
+  pub fn main() {
+    io.println("Hello, world!")
+  }
+  ```
+
+  ([Surya Rose](https://github.com/gearsdatapacks))
+
 ### Bug Fixes
 
 - Fixed a bug where the warnings were printed above the errors without any new line between them.